Проведена большая работа по ТЗ-11, а также были внесены корректировки в файлы ТЗ-10.5 с целью приведения их в соответствие с ТЗ-11

+++ src/main/resources/schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS films (
 	film_description VARCHAR,
 	film_release_date DATE,
 	film_duration INTEGER,
-	film_rate FLOAT,
+	film_rate INTEGER,
 	adult_rate_ID INTEGER REFERENCES adult_rate (adult_rate_ID)
 );
